+++ resources/views/dashboard/layouts/footers/auth.blade.php
@@ -1,0 +1,3 @@
+<footer class="footer">
+    @include('dashboard.layouts.footers.nav')
+</footer>