+++ resources/views/dashboard/layouts/navbars/navbar.blade.php
@@ -1,0 +1,7 @@
+@auth()
+    @include('dashboard.layouts.navbars.navs.auth')
+@endauth
+    
+@guest()
+    @include('dashboard.layouts.navbars.navs.guest')
+@endguest