+++ resources/views/layouts/anime_card.blade.php
@@ -1,6 +1,6 @@
 <div
  class="card card-profile shadow mb-5 anime-card"
- style="background-image: linear-gradient(rgba(94, 114, 228, 0.3), rgba(23, 43, 77, 0.3)), url({{ $anime->cover_url != '' ? $anime->cover_url : $anime->image_url }});color: #fff;border-radius: 10px;margin-top: 0;">
+ style="background-origin: border-box;background-image: linear-gradient(rgba(94, 114, 228, 0.3), rgba(23, 43, 77, 0.3)), url({{ $anime->cover_url != '' ? $anime->cover_url : $anime->image_url }});color: #fff;border-radius: 10px;margin-top: 0;">
     <div class="px-4">
         <div class="row justify-content-center">
             <div class="col-md-6">
@@ -24,7 +24,7 @@
                     @foreach (json_decode($anime->genres) as $gener)
                     <a href="/tags/{{$gener}}" class="btn btn-sm btn-default float-right mx-1">{{$gener}}</a>
                     @endforeach
-                    <a href="/animes/{{$anime->id}}" class="btn btn-sm btn-success float-right mr-4">{{$anime->status}}</a>
+                    <a href="/status/{{$anime->status}}" class="btn btn-sm btn-success float-right mx-1">{{$anime->status}}</a>
                 </div>
             </div>
             
@@ -38,14 +38,14 @@
                                 <h3 style="color: #fff;">{{$anime->title}}</h3>
                             </div>
                             <p
-                            class="p-4"
+                            class="fz-sm-1 text-white p-4"
                             style="background: rgba(23, 43, 77, 0.5);border-radius: 15px;">
                             {{$anime->arabic_synopsis}}
                             </p>
                         </div>
                         <div class="col-md-3 mt-md-5 d-md-none d-lg-block d-sm-none">
                             <a href="animes/{{$anime->id}}">
-                                <img src="{{$anime->image_url}}" class="img-fluid rounded shadow" alt="{{$anime->title}}">
+                                <img src="{{$anime->image_url}}" class="d-sm-hidden img-fluid rounded shadow" alt="{{$anime->title}}">
                             </a>
                         </div>
                     </div>