+++ resources/views/welcome.blade.php
@@ -32,7 +32,7 @@
 
                                 <!-- Links' Buttons -->
                                 <div class="btn-wrapper mt-5">
-                                    <a href="#" class="btn btn-lg btn-github btn-icon mb-3 mb-sm-0" target="_blank">
+                                    <a href="/animes" class="btn btn-lg btn-github btn-icon mb-3 mb-sm-0" target="_blank">
                                         <span class="btn-inner--icon"><i class="fa fa-th-large"></i></span>
                                         <span class="btn-inner--text">قائمة <span class="text-warning">الأنمي</span></span>
                                     </a>
@@ -146,7 +146,7 @@
                     
                     <!-- Section Buttons -->
                     <div class="btn-wrapper">
-                        <a href="#" class="btn btn-primary">مشاهدة الأنمي بحسب التصنيف</a>
+                        <a href="/tags" class="btn btn-primary">مشاهدة الأنمي بحسب التصنيف</a>
                         <a href="https://nucleoapp.com/?ref=1712" target="_blank" class="btn btn-default mt-3 mt-md-0">مشاهدة كل الأستديوهات</a>
                     </div>
                 </div>
@@ -180,7 +180,7 @@
                     
                     <!-- Section Buttons -->
                     <div class="btn-wrapper">
-                        <a href="examples/login.html" class="btn btn-success">إشترك الآن</a>
+                        <a href="/mail" class="btn btn-success">إشترك الآن</a>
                         <a href="examples/register.html" class="btn btn-white">إعدادات بريدك الإلكتروني</a>
                     </div>
                 </div>
@@ -195,13 +195,16 @@
                                     <small>ضع بريدك الإلكتروني وإختر التصنيفات التي تريد أن تصلك آخر تحديثاتها.</small>
                                 </div>
 
-                                <form>
+                                <form method="POST" action="/mail">
+                                    @csrf
                                     <div class="form-group mb-3">
                                         <div class="input-group input-group-alternative">
                                             <div class="input-group-prepend">
                                                 <span class="input-group-text"><i class="ni ni-email-83"></i></span>
                                             </div>
-                                            <input class="form-control" placeholder="البريد الإلكتروني" type="email">
+                                            <input class="form-control" placeholder="البريد الإلكتروني" type="email" name="email">
+                                            <input type="hidden" name="previous_url" value="{{ url()->previous() }}">
+                                            <input type="hidden" name="current_url" value="{{ url()->current() }}">
                                         </div>
                                     </div>
                                     <div class="text-center">