+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# 0.18.0
+## Changed 
+- hide the anime cover on mobiles
+- made the anime synopsis and episodes links smaller
+- made the search ison in /search page smaller on mobiles
+- fixed the background border in anime cards
+- translated broadcast column in animes table
+- filled created_at and updated_at columns in animes table
+
+## Added
+- Installed Argon Dashboard
+- Subscription page and form
+
+----
+
 # 0.17.0
 # Added
 - Google Search page emmbeded in /search page
@@ -11,11 +26,15 @@
 - centered footer icons and links
 - translated all status, rating, broadcast and genres columns
 
+----
+
 # 0.16.2
 ## Changed
 - removed comments from css/app.css and minified it.
 - removed nucleo icons.
 - fixed the background image in anime details page.
+
+----
 
 # 0.16.1
 ## Changed 
@@ -26,6 +45,8 @@
 ## Added
 - Contect page "/contact"
 - changed all assets links to CDN
+
+----
 
 # 0.16.0
 ## Changed
@@ -38,6 +59,7 @@
 ## Added
 - google analytics tag
 
+----
 
 # 0.15.0 - 2020-04-26
 ## Changed